refactor: mutation / search -> making a modal a hook
- modal을 전역적으로 편리하게 재사용하기 위해 커스텀 훅으로 리팩토링
